requirements.txt
apache-airflow-providers-apache-spark
minio
polars
ibis-framework